Move ValueWith* DataTypes to Funcky.DataTypes
