+++ GlobalUsings.props
@@ -2,7 +2,7 @@
 <Project>
 	<ItemGroup Label="Global Usings">
 		<Using Include="System.Diagnostics.Contracts" />
-		<Using Include="Funcky.Extensions" />
+		<Using Include="Funcky.DataTypes" />
 		<Using Include="Funcky.Extensions" />
 		<Using Include="Funcky.Monads" />
 		<Using Include="Funcky.Functional" Static="true" />
